TCF Debugger on Windows now provides UI to control exception handling, similar to signals handling on UNIX

+++ plugins/org.eclipse.tm.tcf/src/org/eclipse/tm/internal/tcf/services/remote/ProcessesProxy.java
@@ -271,7 +271,7 @@ public class ProcessesProxy implements IProcesses {
         }.token;
     }
 
-    public IToken signal(String context_id, int signal, final DoneCommand done) {
+    public IToken signal(String context_id, long signal, final DoneCommand done) {
         return new Command(channel, ProcessesProxy.this,
                 "signal", new Object[]{ context_id, signal }) {
             @Override
